Create Loops_Divisible by 3 or 5 but not by both Series
Write a program to generate a series of numbers between 1 and n (inclusive) that are divisible by either 3 or 5 but not both. The program should take an integer n as input and output the numbers satisfying this condition in a single line, separated by spaces.

Input Format

A single integer n, the upper limit of the range.
Constraints

1 ≤ n ≤ 10,000
Output Format

A single line containing numbers between 1 and n (inclusive) that are divisible by either 3 or 5, but not both, separated by spaces. If no such numbers exist, print nothing.